B/G computation fixes
Properly handle the factors of delta that appear due to the definition
of phi_k for histograms.

Fix overflow in compute_g_...

Rename functions and fix doc to indicate that intermediate functions
compute integer counts and not the B/G value directly.
This integer count must be multiplied by a factor derived from delta in both
case to form the final value.