Add generic algorithm for sum shape(xm-xl)
This algorithm is used to compute B for now, in the general case.
It sums f(x_m-x_l) for x_m/x_l points in a sorted set.

The general algortihm is the original one used before the linear
optimisation for histograms without kernels.

Its complexity is O(|N|^2) in worst case.
It should be closer to O(|N| * density(N) * width(shape)) average.