Rewrite lassoshooting with eigen
This code implements the lassoshooting computation of the R package with
the configuration for our case.

Only parts relevant to our computation have been kept.
The rest of the code has been simplified away.

This is not a direct reuse of the code.
The original code used cblas/fortran blas from R headers.
The rewrite uses only eigen.

This has not been tested yet.