Second phase of process input
After parsing interval sets from files, convert them to points, do
sorting and reversing (if requested), put the point sets into a simple
vector2d(process,region).

The vector2d is set so that rows are regions, and all point sets for a
region can be retrieved as a span<...>.
This simplifies the api of computation functions, which now operate on a
list of point sets (for each process), and can forget about the notion
of region for most of them.