Add main selector for base/kernel variants
Change code so that combinations of base and kernels is encoded at the
level of compute_b/g/b_hat.

compute_intermediate_values is templated on base and kernel:
- correct case will just apply the compute_* functions on all regions
- bad case (base = None) will throw.

This way the selection of base/kernel combinations is done with a single
visit(do_stuff, base, kernels).

For now, the kernel part is not selected upon.
TODO: compute b_hat approx, and add kernel selection.