Add asserts to constructor of structs with restricted domains
Add asserts to check in debug mode that many structs have correct
values:
- shapes
- bases
- kernels

The code does not support an interval kernel of width 0, which was the
cause of the Nans detected in B.
The normalization factor is the cause of Nans.
Current fix is to restrict the interval kernel to width > 0.

Thus all widths are required to be > 0:
- deduced ones are rounded to 1
- explicit are checked to be >0