Fixes and more debug
Fixed degenerate cases due to kernel_width < 2.
This caused the G matrix to contains lots of 0s.

Now lassoshooting diverges in some cases:
- prdm9 example, K=4, delta=10k, kernels => diverges to +inf
- prdm9 example, K=4, delta=10k, no_kernel => converges.
- prdm9 example, K=4, delta=1k, kernels => converges.
- prdm9 example, K=5, delta=1k, kernels => converges to values ~10^80.

Testing showed that divergence also occurs in the R code, so the rewrite is not the cause.
More testing indicated that the penaltyweights array is working as intended (compared to adapting the values like in the old code).

In all cases, the G^-1 * B non-sparse solution exists and does not look crazy.

G, B and D all have huge values (~10^[7 to 9]), but they should be normalized correctly.

Added more diagnostics.
Use old lambda = 2, with FIXME to question the choice.
Use gamma = 1 by default instead of 3.
(both gamma and lambda do not make much difference in the above problem).