Remove PointInNZD optimization
This optimization tag was useful to remove some comparisons in the
generated code.
With the shift to float pointspace, the property represented by the tag
was violated by a small margin in some cases (10^-13).
This is expected from the floating point space.
Instead of trying to fix it, just remove the optimization completely.
The performance impact is in fact negligible.