Use add to simplify conv(trapezoids) in computations
The code using shapes is greatly simplified.
This is also faster in practice, because we sum over points only once.
Results are equivalents.

Order of approximations has been changed in the b_hat computation.
This reduces the over-approximation in practice, changing the D and lasso
values a bit (but not G, B or G^-1B).