Reorganize shape::combinators
Group combinator factory functions with their base type.
The initial split was for using priorities in them, which is not
required.

Combinator transformation: split the basic ones from convolution which
uses priorities.
Start with the basic ones (component).
Then introduce the priorities, with doc for the enable_if part.
Add constexpr values for the priority levels, this is easier to read
than magic numbers in enable_if statements.