Eigen: use pre-downloaded version
Include the latest version of eigen to the repository as a tar.gz.
This avoids potential failures due to:
- missing curl command
- no internet connection
- URL changes for eigen download

The version was fixed anyway, so the only inconvenient is added data to the git.
This should make the code "last" longer (be usable longer, as there are less dependencies and way to fail the compilation part).