Precompute binomial coefficients
The number required is fixed at compilation (number of nested convolution and degree of base shapes).
Precomputing it should be more efficient.
In the unlikely case where N=20 is not sufficient, the computation will probably be very slow as degrees would be on the order of 10 and computation is O(degree^4)