make: properly cache untar-ed eigen
Handling of bundled eigen lib is somewhat hacky.
It relies on untar-ed archive dir as a target.
The new version has a dir which is "old" (in the tar archive).
As make relies on date, it is always resfreshed by un-taring again.
Fixed by updating dir date with touch command.